Add notes expressing doubt about encoding
I don't think we have any categorical features. Everything seems to have
a ranking. E.g., the presence of cavities is worse than the absence. And
our sex value is a float where the lower value is more likely to be
female and higher values are more likely to be male. We don't actually
know if the subject is a male or female. So these all really seem to be
ordinal.